# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: http://localhost:3001/
...
paths:
  "/api":
    get:
      tags:
      - api
      summary: Retrieve the Film Manager
      description: The Film Manager resource, representing the entry point of the
        REST interface, with ID filmId is retrieved. This operation does not require
        authentication.
      operationId: getFilmManager
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FilmManager"
        '500':
          description: Internal Server Error
          content: {}
  "/api/films":
    post:
      tags:
      - api/films
      security:
      - cookieAuth: []
      summary: Create a new film
      description: A new film is created by the authenticated user (who becomes the
        owner).
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created (with no id because
          it is assigned by the service)
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/private":
    get:
      tags:
      - api/films/private
      security:
      - cookieAuth: []
      summary: Retrieve the private films of the logged-in user
      description: The private films of the logged-in user are retrieved. A pagination
        mechanism is used to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/private/{filmId}":
    get:
      tags:
      - api/films/private/{filmId}
      security:
      - cookieAuth: []
      summary: Retrieve a private film
      description: The private film with ID filmId is retrieved. This operation can
        be performed on the film if the user who performs the operation is the film's
        owner.
      operationId: getSinglePrivateFilm
      parameters:
      - name: filmId
        in: path
        description: ID of the film to retrieve
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
      - api/films/private/{filmId}
      security:
      - cookieAuth: []
      summary: Update a private film
      description: The private film with ID filmId is updated. This operation does
        not allow changing its visibility. This operation can be performed only by
        the owner.
      operationId: updateSinglePrivateFilm
      parameters:
      - name: filmId
        in: path
        description: ID of the film to update
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
      - api/films/private/{filmId}
      security:
      - cookieAuth: []
      summary: Delete a private film
      description: The private film with ID filmId is deleted. This operation can
        only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
      - name: filmId
        in: path
        description: ID of the film to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public":
    get:
      tags:
      - api/films/public
      summary: Retrieve the public films
      description: The public films (i.e., the films that are visible for all the
        users of the service) are retrieved. This operation does not require authentication.
        A pagination mechanism is used to limit the size of messages.
      operationId: getPublicFilms
      parameters:
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/invited":
    get:
      tags:
      - api/films/public/invited
      security:
      - cookieAuth: []
      summary: Retrieve the public films that the logged-in user has been invited
        to review
      description: The public films that the logged-in user has been invited to review
        are retrieved. A pagination mechanism is used to limit the size of messages.
        The user can filter by status to avoid unnecessary information.
      operationId: getInvitedFilms
      parameters:
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: The id of the requested page (if absent, the first page is returned)
      - in: query
        name: status
        required: false
        schema:
          type: string
          enum: [pending, accepted, cancelled, completed]
        description: Filter invitations by their current status
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/invited/acceptance":
    post:
      tags:
      - api/films/public/invited
      security:
      - cookieAuth: []
      summary: Accept all pending invitations
      description: Allows the logged-in user to accept all pending review invitations
        in a single operation.
      operationId: acceptAllInvitations
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/{filmId}":
    get:
      tags:
      - api/films/public/{filmId}
      summary: Retrieve a public film
      description: The public film with ID filmId is retrieved. This operation does
        not require authentication.
      operationId: getSinglePublicFilm
      parameters:
      - name: filmId
        in: path
        description: ID of the film to retrieve
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
      - api/films/public/{filmId}
      security:
      - cookieAuth: []
      summary: Update a public film
      description: The public film with ID filmId is updated. This operation does
        not allow changing its visibility. This operation can be performed only by
        the owner.
      operationId: updateSinglePublicFilm
      parameters:
      - name: filmId
        in: path
        description: ID of the film to update
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
      - api/films/public/{filmId}
      security:
      - cookieAuth: []
      summary: Delete a public film
      description: The public film with ID filmId is deleted. This operation can only
        be performed by the owner.
      operationId: deleteSinglePublicFilm
      parameters:
      - name: filmId
        in: path
        description: ID of the film to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/{filmId}/reviews":
    post:
      tags:
      - api/films/public/{filmId}/reviews
      security:
      - cookieAuth: []
      summary: Issue film review to some users
      description: The film with ID filmId is assigned to one or more users for review
        and the corresponding reviews are created. The users are specified in the
        review representations in the request body. This operation can only be performed
        by the owner.
      operationId: issueFilmReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: the new film reviews, including the users to whom they are issued
        content:
          application/json:
            schema:
              type: array
              items:
                "$ref": "#/components/schemas/Review"
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Review"
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    get:
      tags:
      - api/films/public/{filmId}/reviews
      summary: Retrieve the list of all the reviews that have been issued/completed
        for a public film
      description: 'All the reviews that have been issued/completed for the public
        film with ID filmId are retrieved. A pagination mechanism is used to limit
        the size of messages. The owner can filter by status.'
      operationId: getFilmReviews
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose reviews must be retrieved
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: ID of the requested page (if absent, the first page is returned)
      - in: query
        name: status
        required: false
        schema:
          type: string
          enum: [pending, accepted, cancelled, completed]
        description: Filter invitations by their current status
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Reviews"
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/{filmId}/reviews/{reviewerId}":
    get:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}
      summary: Retrieve a review that has been issued/completed for a film
      description: 'The review of the film with ID filmID issued to the user with
        ID reviewerId is retrieved. This operation does not require authentication. '
      operationId: getSingleReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose reviews must be retrieved
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the user to whom the review has been issued
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Review"
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}
      security:
      - cookieAuth: []
      summary: Complete a review
      description: The review of the film with ID filmId and issued to the user with
        ID reviewerId is completed. This operation only allows setting the "completed"
        property to the "true" value, and changing the values of the "reviewDate",
        "rating", and "review" properties. This operation can be performed only by
        the invited reviewer.
      operationId: updateSingleReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose review must be completed
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the user to whom the review has been issued
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: The updated Review object
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Review"
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (the user is not the reviewer)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}
      security:
      - cookieAuth: []
      summary: Delete a review invitation
      description: The review of the film with ID filmId and issued to the user with
        ID reviewerId is deleted. This operation can only be performed by the owner,
        and only if the review has not yet been completed by the reviewer.
      operationId: deleteSingleReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose review invitation must be deleted
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the user to whom the review has been issued
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (the user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (the review has already been completed)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/assignments":
    post:
      tags:
      - api/films/public/assignments
      security:
      - cookieAuth: []
      summary: Assign automatically review invitations for the unassigned public films
        owned by the authenticated user in a balanced manner
      description: The films owned by the authenticated user who performed the request
        and that are not assigned to any user for review, are automatically assigned
        to the users of the service in a balanced manner. The assignments are returned.
        Only authenticated users can make this request.
      operationId: assignReviewBalanced
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    filmId:
                      type: integer
                    reviewerId:
                      type: integer
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/users":
    get:
      tags:
      - api/users
      security:
      - cookieAuth: []
      summary: Get information about the users
      description: The available information (passwords excluded) about all the users
        is retrieved. This operation is available only to authenticated users.
      operationId: getUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not Found
          content: {}
  "/api/users/{userId}":
    get:
      tags:
      - api/users/{userId}
      security:
      - cookieAuth: []
      summary: Get information about a user
      description: The available information (password excluded) about the user specified
        by userId is retrieved. This operation requires authentication.
      operationId: getSingleUser
      parameters:
      - name: userId
        in: path
        description: ID of the user to get
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not Found
          content: {}
  "/api/users/authenticator":
    post:
      tags:
      - api/users/authenticator
      summary: Logs a user in
      description: The user who wants to log in sends the user data to the
        authenticator which performs the operation. If the request for the login of a new user
        comes from an already authenticated user, the previous user is first logged out.
      operationId: authenticateUser
      requestBody:
        description: The data of the user who wants to perform log in. The data
          structure must contain email and password.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/User"
        required: true
      responses:
        '200':
          description: Successful operation
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '401':
          description: Unauthorized
          content: {}
  "/api/users/authenticator/current":
    delete:
      tags:
        - "api/users/authenticator/current"
      summary: "Logs the current user out"
      description:
        Invalidates the current user session.  
        Removes the authentication cookie if present.
      operationId: "logoutUser"
      responses:
        "204":
          description: "No Content"
        "401":
          description: "Unauthorized"
components:
  schemas:
    FilmManager:
      description: schema for FilmManager data structure (api entry point)
      type: object
      additionalProperties: false
      properties:
        films:
          type: string
          description: URI to access the api/films endpoint
          format: uri
        privateFilms:
          type: string
          description: URI to access the api/films/private endpoint
          format: uri
        publicFilms:
          type: string
          description: URI to access the api/films/public endpoint
          format: uri
        invitedPublicFilms:
          type: string
          format: uri
          description: URI to access the api/films/public/invited endpoint
        reviewAssignments:
          type: string
          description: URI to access the api/films/public/assignments endpoint
          format: uri
        users:
          type: string
          description: URI to access the api/users endpoint
          format: uri
        usersAuthenticator:
          type: string
          description: URI to access the api/users/authenticator endpoint
          format: uri
...
    Users:
      description: schema for Users data structures
      type: array
      items:
        "$ref": "#/components/schemas/User"
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access
            it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite ones of the owner,
            false otherwise '
          default: 'false'
        self:
          type: string
          format: uri
        reviews:
          type: string
          format: uri
          description: the reviews for the film
      required:
      - title
      - owner
      - private
    Films:
      description: schema for Films data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the films
        totalPages:
          type: integer
          minimum: 0
          description: total number of pages
        currentPage:
          type: integer
          minimum: 0
          description: the page included in this object
        totalItems:
          type: integer
          minimum: 0
          description: the total number of items in the current page
        films:
          type: array
          items:
            "$ref": "#/components/schemas/Film"
          description: the films of the current page
        next:
          type: string
          format: uri
      required:
      - totalPages
    Review:
      description: schema for Review data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        filmId:
          type: integer
          description: unique identifier of the film for which a review invitation
            has been issued
        reviewerId:
          type: integer
          description: unique identifier of the user who has received the review invitation
        completed:
          type: boolean
          description: true if the review has been completed, false otherwise
          default: 'false'
        reviewDate:
          type: string
          format: date
          description: the date when the review has been completed by the invited
            user
        rating:
          type: integer
          description: the rating the reviewer has given to the film
          minimum: 1
          maximum: 10
        review:
          type: string
          description: a textual description of the review
          maxLength: 1000
        invitationStatus:
          type: string
          description: The current status of the review invitation
          enum: [pending, accepted, cancelled, completed]
        expirationDate:
          type: string
          format: date-time
          description: The date and time when the invitation expires
        self:
          type: string
          format: uri
      required:
      - filmId
      - reviewerId
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid